refactor(ud-webhook): merge alpha adapter helpers into validation test
- Move helper fns from uniteddeployment_validation_alpha_adapter_test.go
  into uniteddeployment_validation_test.go; file had no test cases
- Remove unused validateUnitedDeploymentSpec
- Fix copyright year 2020→2026 in uniteddeployment_types.go

Refs furykerry review comments
